+++ src/law.rs
@@ -15,15 +15,15 @@ impl ops::Div<Resistance> for Voltage {
 
         let micro_volts = self.micro_volts().unsigned_abs();
 
-        let nano_volts = (micro_volts as u64)
-            .checked_mul(1000u64)
+        let nano_volts = micro_volts
+            .checked_mul(1_000u64)
             .expect("Voltage would overflow");
 
         let micro_amps = nano_volts
             .checked_div(resistance.milli_ohms() as u64)
             .expect("Current would overflow");
 
-        Current::from_micro_amps(micro_amps as u32)
+        Current::from_micro_amps(micro_amps as u64)
     }
 }
 
@@ -34,17 +34,17 @@ impl ops::Mul<Resistance> for Current {
     ///
     /// Will be rounded down to the nearest whole microvolt (μV).
     fn mul(self, resistance: Resistance) -> Self::Output {
-        let micro_amps = self.micro_amps() as u64;
+        let micro_amps = self.micro_amps();
 
         let nano_volts = micro_amps
-            .checked_mul(resistance.milli_ohms() as u64)
+            .checked_mul(resistance.milli_ohms())
             .expect("Voltage would overflow");
 
         let micro_volts = nano_volts
-            .checked_div(1000u64)
+            .checked_div(1_000u64)
             .expect("Voltage would overflow");
 
-        Voltage::from_micro_volts(micro_volts as i32)
+        Voltage::from_micro_volts(micro_volts as i64)
     }
 }
 
@@ -73,15 +73,15 @@ impl ops::Div<Current> for Voltage {
 
         let micro_volts = self.micro_volts().unsigned_abs();
 
-        let nano_volts = (micro_volts as u64)
-            .checked_mul(1000u64)
+        let nano_volts = micro_volts
+            .checked_mul(1_000u64)
             .expect("Voltage would overflow");
 
         let milli_ohms = nano_volts
-            .checked_div(current.micro_amps() as u64)
+            .checked_div(current.micro_amps())
             .expect("Resistance would overflow");
 
-        Resistance::from_milli_ohms(milli_ohms as u32)
+        Resistance::from_milli_ohms(milli_ohms)
     }
 }
 
@@ -94,9 +94,9 @@ mod tests {
     #[test_case(-5_000_000, 100_000, 50_000; "negative 5V, 100Ω equals 50,000μA")]
     #[test_case(3_300_000, 4_700_000, 702; "positive 3.3V, 4.7kΩ equals 702μA")]
     fn test_current_equals_voltage_over_resistance(
-        micro_volts: i32,
-        milli_ohms: u32,
-        expected_micro_amps: u32,
+        micro_volts: i64,
+        milli_ohms: u64,
+        expected_micro_amps: u64,
     ) {
         let v = Voltage::from_micro_volts(micro_volts);
         let r = Resistance::from_milli_ohms(milli_ohms);
@@ -108,9 +108,9 @@ mod tests {
     #[test_case(25_000, 75_000, 1_875_000; "25mA, 75Ω equals 1_875_000μV")]
     #[test_case(39_000, 162_000, 6_318_000; "39mA, 162Ω equals 6_318_000μV")]
     fn test_voltage_equals_current_times_resistance(
-        micro_amps: u32,
-        milli_ohms: u32,
-        expected_micro_volts: i32,
+        micro_amps: u64,
+        milli_ohms: u64,
+        expected_micro_volts: i64,
     ) {
         let i = Current::from_micro_amps(micro_amps);
         let r = Resistance::from_milli_ohms(milli_ohms);
@@ -122,9 +122,9 @@ mod tests {
     #[test_case(25_000, 75_000, 1_875_000; "25mA, 75Ω equals 1_875_000μV")]
     #[test_case(39_000, 162_000, 6_318_000; "39mA, 162Ω equals 6_318_000μV")]
     fn test_voltage_equals_resistance_times_current(
-        micro_amps: u32,
-        milli_ohms: u32,
-        expected_micro_volts: i32,
+        micro_amps: u64,
+        milli_ohms: u64,
+        expected_micro_volts: i64,
     ) {
         let r = Resistance::from_milli_ohms(milli_ohms);
         let i = Current::from_micro_amps(micro_amps);
@@ -138,9 +138,9 @@ mod tests {
     #[test_case(6_318_000, 39_000, 162_000; "positive 6.318V, 39mA equals 162Ω")]
     #[test_case(-6_318_000, 39_000, 162_000; "negative 6.318V, 39mA equals 162Ω")]
     fn test_resistance_equals_voltage_over_current(
-        micro_volts: i32,
-        micro_amps: u32,
-        expected_milli_ohms: u32,
+        micro_volts: i64,
+        micro_amps: u64,
+        expected_milli_ohms: u64,
     ) {
         let v = Voltage::from_micro_volts(micro_volts);
         let i = Current::from_micro_amps(micro_amps);