+++ src/prelude.rs
@@ -1,4 +1,10 @@
-pub use crate::current::{Current, ExtF32 as CurrentExtF32, ExtU32 as CurrentExtU32};
+pub use crate::current::{
+    Current, FromFloat as CurrentFromFloat, FromInteger as CurrentFromInteger,
+};
 pub use crate::law::*;
-pub use crate::resistance::{ExtF32 as ResistanceExtF32, ExtU32 as ResistanceExtU32, Resistance};
-pub use crate::voltage::{ExtF32 as VoltageExtF32, ExtI32 as VoltageExtI32, Voltage};
+pub use crate::resistance::{
+    FromFloat as ResistanceFromFloat, FromInteger as ResistanceFromInteger, Resistance,
+};
+pub use crate::voltage::{
+    FromFloat as VoltageFromFloat, FromInteger as VoltageFromInteger, Voltage,
+};