+++ src/resistance.rs
@@ -1,7 +1,7 @@
-use crate::{assert, helpers};
+use crate::assert_positive_float;
 use core::{cmp, ops};
 
-/// Represents a resistance value, stored as whole milliohms (mΩ).
+/// Represents a resistance value, stored as whole milliohms (mΩ) stored in a `u64` value.
 /// This value can only be positive.
 ///
 /// **Reminder:** `1000 mΩ = 1 Ω, 1000 Ω = 1 kΩ, 1000 kΩ = 1 MΩ`
@@ -10,7 +10,7 @@ use core::{cmp, ops};
 ///
 /// # Creating a Resistance value
 /// You can create a `Resistance` value using the `from_milli_ohms` method, or using one of the
-/// extension methods on `u32` and `f32`:
+/// extension methods on integer and floating-point types:
 ///
 /// ```rust
 /// use ohms::prelude::*;
@@ -19,9 +19,9 @@ use core::{cmp, ops};
 ///
 /// // More ergonomic:
 /// let r2 = 100.milli_ohms(); // 0.1Ω
-/// let r3 = 220u32.ohms(); // 220Ω
-/// let r4 = 1.5f32.kilo_ohms(); // 1.5kΩ
-/// let r5 = 1.5f32.mega_ohms(); // 1.5MΩ
+/// let r3 = 220.ohms(); // 220Ω
+/// let r4 = 1.5.kilo_ohms(); // 1.5kΩ
+/// let r5 = 1.5.mega_ohms(); // 1.5MΩ
 /// ```
 ///
 /// # Comparing Resistance values
@@ -30,8 +30,8 @@ use core::{cmp, ops};
 /// ```rust
 /// use ohms::prelude::*;
 ///
-/// let r1 = 220u32.ohms(); // 220Ω
-/// let r2 = 4.7f32.kilo_ohms(); // 4.7kΩ
+/// let r1 = 220.ohms(); // 220Ω
+/// let r2 = 4.7.kilo_ohms(); // 4.7kΩ
 ///
 /// if r1 > r2 {
 ///     println!("{:?} is greater than {:?}", r1, r2);
@@ -44,91 +44,91 @@ use core::{cmp, ops};
 /// You can use the `+` and `-` operators to add and subtract `Resistance` values from each other.
 /// The result is a new `Resistance` value, rounded down to the nearest whole milliohm (mΩ).
 ///
-/// If the result of the operation would overflow or underflow the `u32` value, the operation will panic.
+/// If the result of the operation would overflow or underflow, the operation will panic.
 ///
 /// ```rust
 /// use ohms::prelude::*;
 ///
-/// let r1 = 220u32.ohms(); // 220Ω
-/// let r2 = 4.7f32.kilo_ohms(); // 4.7kΩ
+/// let r1 = 220.ohms(); // 220Ω
+/// let r2 = 4.7.kilo_ohms(); // 4.7kΩ
 ///
 /// let r3 = r1 + r2; // 4.92kΩ
-/// let r4 = r2 - 2f32.kilo_ohms(); // 2.7kΩ
+/// let r4 = r2 - 2.kilo_ohms(); // 2.7kΩ
 /// ```
 ///
 /// # Scaling Resistance values
-/// You can use the `*` and `/` operators to scale `Resistance` values by a scalar `u32` or `f32` value.
+/// You can use the `*` and `/` operators to scale `Resistance` values by an integer or floating-point value.
 /// The result is a new `Resistance` value, rounded down to the nearest whole milliohm (mΩ).
 ///
-/// If the result of operation would overflow or underflow the `u32` value, the operation will panic.
+/// If the result of operation would overflow or underflow, the operation will panic.
 ///
 /// If the result of the operation would be infinite or NaN, the operation will panic.
 ///
 /// ```rust
 /// use ohms::prelude::*;
 ///
-/// let r1 = 220u32.ohms(); // 220Ω
+/// let r1 = 220.ohms(); // 220Ω
 /// let r2 = r1 * 3; // 660Ω
 ///
-/// let r3 = 47u32.kilo_ohms(); // 47kΩ
-/// let r4 = r3 / 2.5f32; // 18.8kΩ
+/// let r3 = 47.kilo_ohms(); // 47kΩ
+/// let r4 = r3 / 2.5; // 18.8kΩ
 /// ```
 ///
 /// # Converting to other denominations
 /// You can use the `milli_ohms`, `ohms`, `kilo_ohms` and `mega_ohms` methods to convert a `Resistance`
-/// value to a `u32` or `f32` value in the specified denomination.
+/// value to a numeric value in the specified denomination.
 ///
 /// ```rust
 /// use ohms::prelude::*;
 ///
-/// let r1 = 47.5f32.ohms(); // 47.5Ω
+/// let r1 = 47.5.ohms(); // 47.5Ω
 ///
 /// println!("{:.3}kΩ is {:.1}Ω", r1.kilo_ohms(), r1.ohms());
 /// ```
 ///
 #[derive(Clone, Copy, Debug)]
 pub struct Resistance {
-    milliohms: u32,
+    raw: u64,
 }
 
 impl Resistance {
     /// Creates a new `Resistance` from a number of whole milliohms (mΩ).
     ///
     /// It is recommended to use the `milli_ohms`, `ohms`, `kilo_ohms` and `mega_ohms` extension
-    /// methods on `u32` and `f32` instead of this method for ergonomics.
+    /// methods on integer and floating-point values instead.
     #[inline]
-    pub const fn from_milli_ohms(milliohms: u32) -> Resistance {
-        Resistance { milliohms }
+    pub const fn from_milli_ohms(value: u64) -> Resistance {
+        Resistance { raw: value }
     }
 
     /// Returns the resistance value in whole milliohms (mΩ).
     #[inline]
-    pub const fn milli_ohms(&self) -> u32 {
-        self.milliohms
+    pub const fn milli_ohms(&self) -> u64 {
+        self.raw
     }
 
     /// Returns the resistance value in fractional ohms (Ω).
     #[inline]
-    pub fn ohms(&self) -> f32 {
-        self.milliohms as f32 / 1_000f32
+    pub fn ohms(&self) -> f64 {
+        self.raw as f64 / 1_000f64
     }
 
     /// Returns the resistance value in fractional kilohms (kΩ).
     #[inline]
-    pub fn kilo_ohms(&self) -> f32 {
-        self.milliohms as f32 / 1_000_000f32
+    pub fn kilo_ohms(&self) -> f64 {
+        self.raw as f64 / 1_000_000f64
     }
 
     /// Returns the resistance value in fractional megaohms (MΩ).
     #[inline]
-    pub fn mega_ohms(&self) -> f32 {
-        self.milliohms as f32 / 1_000_000_000f32
+    pub fn mega_ohms(&self) -> f64 {
+        self.raw as f64 / 1_000_000_000f64
     }
 
     /// Returns whether the resistance value is zero ohms (0Ω).
     #[inline]
     pub const fn is_zero(&self) -> bool {
-        self.milliohms == 0
+        self.raw == 0
     }
 
     /// Returns a `Resistance` value of zero ohms (0Ω).
@@ -138,39 +138,37 @@ impl Resistance {
     }
 }
 
-// Equality traits
 impl PartialEq for Resistance {
     #[inline]
     fn eq(&self, other: &Resistance) -> bool {
-        self.milliohms == other.milliohms
+        self.raw == other.raw
     }
 }
 
 impl Eq for Resistance {}
 
-// Comparison traits
 impl PartialOrd for Resistance {
     #[inline]
     fn partial_cmp(&self, other: &Resistance) -> Option<cmp::Ordering> {
-        self.milliohms.partial_cmp(&other.milliohms)
+        self.raw.partial_cmp(&other.raw)
     }
 }
 
 impl Ord for Resistance {
     #[inline]
     fn cmp(&self, other: &Resistance) -> cmp::Ordering {
-        self.milliohms.cmp(&other.milliohms)
+        self.raw.cmp(&other.raw)
     }
 }
 
-// Math operators
 impl ops::Add for Resistance {
     type Output = Resistance;
 
+    /// Adds two `Resistance` values together, returning a new `Resistance` value.
     #[inline]
     fn add(self, other: Resistance) -> Resistance {
-        self.milliohms
-            .checked_add(other.milliohms)
+        self.raw
+            .checked_add(other.raw)
             .map(Resistance::from_milli_ohms)
             .expect("Overflow when adding resistance values")
     }
@@ -179,86 +177,143 @@ impl ops::Add for Resistance {
 impl ops::Sub for Resistance {
     type Output = Resistance;
 
+    /// Subtracts one `Resistance` value from another, returning a new `Resistance` value.
     #[inline]
     fn sub(self, other: Resistance) -> Resistance {
-        self.milliohms
-            .checked_sub(other.milliohms)
+        self.raw
+            .checked_sub(other.raw)
             .map(Resistance::from_milli_ohms)
             .expect("Overflow when subtracting resistance values")
     }
 }
 
-impl ops::Mul<u32> for Resistance {
-    type Output = Resistance;
+macro_rules! impl_mul_for_integer {
+    ($i:ty) => {
+        impl ops::Mul<$i> for Resistance {
+            type Output = Resistance;
 
-    #[inline]
-    fn mul(self, other: u32) -> Resistance {
-        self.milliohms
-            .checked_mul(other)
-            .map(Resistance::from_milli_ohms)
-            .expect("Overflow when multiplying resistance value")
-    }
+            /// Multiplies a `Resistance` value by an integer value, returning a new `Resistance` value.
+            #[inline]
+            #[allow(unused_comparisons)]
+            fn mul(self, scale_factor: $i) -> Resistance {
+                if scale_factor < 0 {
+                    panic!("Cannot multiply resistance value by negative value");
+                }
+                self.raw
+                    .checked_mul(scale_factor as u64)
+                    .map(Resistance::from_milli_ohms)
+                    .expect("Overflow when multiplying resistance value")
+            }
+        }
+    };
 }
+
+impl_mul_for_integer!(u8);
+impl_mul_for_integer!(u16);
+impl_mul_for_integer!(u32);
+impl_mul_for_integer!(u64);
+impl_mul_for_integer!(i8);
+impl_mul_for_integer!(i16);
+impl_mul_for_integer!(i32);
+impl_mul_for_integer!(i64);
 
 impl ops::Mul<f32> for Resistance {
     type Output = Resistance;
 
+    /// Multiplies the `Resistance` value by a floating-point value, returning a new `Resistance` value.
     #[inline]
-    fn mul(self, other: f32) -> Resistance {
-        let milliohms = match other {
-            _ if other.is_infinite() => {
-                panic!("Cannot multiply resistance value by infinity")
-            }
-            _ if other.is_nan() => panic!("Cannot multiply resistance value by NaN"),
-            _ if other.is_sign_negative() => {
-                panic!("Cannot multiply resistance value by negative value")
-            }
-            _ if other == 0f32 => Some(0),
-            _ => helpers::checked_mul_unsigned_f32(self.milliohms, other),
-        };
-
-        Resistance::from_milli_ohms(milliohms.unwrap())
+    fn mul(self, scale_factor: f32) -> Resistance {
+        self * scale_factor as f64
     }
 }
 
-impl ops::Div<u32> for Resistance {
+impl ops::Mul<f64> for Resistance {
     type Output = Resistance;
 
+    /// Multiplies a `Resistance` value by a floating-point value, returning a new `Resistance` value.
     #[inline]
-    fn div(self, other: u32) -> Resistance {
-        if other == 0 {
-            panic!("Cannot divide resistance value by zero");
-        }
-        self.milliohms
-            .checked_div(other)
-            .map(Resistance::from_milli_ohms)
-            .expect("Overflow when dividing resistance value")
+    fn mul(self, scale_factor: f64) -> Resistance {
+        let result = match scale_factor {
+            _ if scale_factor.is_infinite() => {
+                panic!("Cannot multiply resistance value by infinity")
+            }
+            _ if scale_factor.is_nan() => panic!("Cannot multiply resistance value by NaN"),
+            _ if scale_factor.is_sign_negative() => {
+                panic!("Cannot multiply resistance value by negative value")
+            }
+            _ => self.raw as f64 * scale_factor,
+        };
+
+        Resistance::from_milli_ohms(result as u64)
     }
 }
+
+macro_rules! impl_div_for_integer {
+    ($i:ty) => {
+        impl ops::Div<$i> for Resistance {
+            type Output = Resistance;
+
+            /// Divides a `Resistance` value by an integer value, returning a new `Resistance` value.
+            #[inline]
+            #[allow(unused_comparisons)]
+            fn div(self, divisor: $i) -> Resistance {
+                if divisor == 0 {
+                    panic!("Cannot divide resistance value by zero");
+                } else if divisor < 0 {
+                    panic!("Cannot divide resistance value by negative value");
+                }
+                self.raw
+                    .checked_div(divisor as u64)
+                    .map(Resistance::from_milli_ohms)
+                    .expect("Overflow when dividing resistance value")
+            }
+        }
+    };
+}
+
+impl_div_for_integer!(u8);
+impl_div_for_integer!(u16);
+impl_div_for_integer!(u32);
+impl_div_for_integer!(u64);
+impl_div_for_integer!(i8);
+impl_div_for_integer!(i16);
+impl_div_for_integer!(i32);
+impl_div_for_integer!(i64);
 
 impl ops::Div<f32> for Resistance {
     type Output = Resistance;
 
+    /// Divides a `Resistance` value by a floating-point value, returning a new `Resistance` value.
     #[inline]
-    fn div(self, other: f32) -> Resistance {
-        let milliohms = match other {
-            _ if other == 0f32 => panic!("Cannot divide resistance value by zero"),
-            _ if other.is_infinite() => {
-                panic!("Cannot divide resistance value by infinity")
-            }
-            _ if other.is_nan() => panic!("Cannot divide resistance value by NaN"),
-            _ if other.is_sign_negative() => {
-                panic!("Cannot divide resistance value by negative value")
-            }
-            _ => helpers::checked_div_unsigned_f32(self.milliohms, other),
-        };
-
-        Resistance::from_milli_ohms(milliohms.unwrap())
+    fn div(self, divisor: f32) -> Resistance {
+        self / divisor as u64
     }
 }
 
-/// Extension trait for simple short-hands for creating `Resistance` values from `u32` values.
-pub trait ExtU32 {
+impl ops::Div<f64> for Resistance {
+    type Output = Resistance;
+
+    /// Divides a `Resistance` value by a floating-point value, returning a new `Resistance` value.
+    #[inline]
+    fn div(self, divisor: f64) -> Resistance {
+        let result = match divisor {
+            _ if divisor == 0f64 => panic!("Cannot divide resistance value by zero"),
+            _ if divisor.is_infinite() => {
+                panic!("Cannot divide resistance value by infinity")
+            }
+            _ if divisor.is_nan() => panic!("Cannot divide resistance value by NaN"),
+            _ if divisor.is_sign_negative() => {
+                panic!("Cannot divide resistance value by negative value")
+            }
+            _ => (self.raw as f64) / divisor,
+        };
+
+        Resistance::from_milli_ohms(result as u64)
+    }
+}
+
+/// Extension trait for simple short-hands for creating `Resistance` values from integer values.
+pub trait FromInteger {
     /// Creates a new `Resistance` from a number of whole milliohms (mΩ).
     fn milli_ohms(self) -> Resistance;
 
@@ -272,388 +327,97 @@ pub trait ExtU32 {
     fn mega_ohms(self) -> Resistance;
 }
 
-impl ExtU32 for u32 {
-    #[inline]
-    fn milli_ohms(self) -> Resistance {
-        Resistance::from_milli_ohms(self)
-    }
+macro_rules! impl_resistance_from_integer {
+    ($i:ty) => {
+        impl FromInteger for $i {
+            #[inline]
+            fn milli_ohms(self) -> Resistance {
+                Resistance::from_milli_ohms(self as u64)
+            }
 
-    #[inline]
-    fn ohms(self) -> Resistance {
-        let milliohms = self
-            .checked_mul(1_000)
-            .expect("Overflow when converting ohms to milliohms");
-        Resistance::from_milli_ohms(milliohms)
-    }
+            #[inline]
+            fn ohms(self) -> Resistance {
+                let milliohms = (self as u64)
+                    .checked_mul(1_000)
+                    .expect("Overflow when converting ohms to milliohms");
+                Resistance::from_milli_ohms(milliohms)
+            }
 
-    #[inline]
-    fn kilo_ohms(self) -> Resistance {
-        let milliohms = self
-            .checked_mul(1_000_000)
-            .expect("Overflow when converting kilohms to milliohms");
-        Resistance::from_milli_ohms(milliohms)
-    }
+            #[inline]
+            fn kilo_ohms(self) -> Resistance {
+                let milliohms = (self as u64)
+                    .checked_mul(1_000_000)
+                    .expect("Overflow when converting kilohms to milliohms");
+                Resistance::from_milli_ohms(milliohms)
+            }
 
-    #[inline]
-    fn mega_ohms(self) -> Resistance {
-        let milliohms = self
-            .checked_mul(1_000_000_000)
-            .expect("Overflow when converting megaohms to milliohms");
-        Resistance::from_milli_ohms(milliohms)
-    }
+            #[inline]
+            fn mega_ohms(self) -> Resistance {
+                let milliohms = (self as u64)
+                    .checked_mul(1_000_000_000)
+                    .expect("Overflow when converting megaohms to milliohms");
+                Resistance::from_milli_ohms(milliohms)
+            }
+        }
+    };
 }
 
-/// Extension trait for simple short-hands for creating `Resistance` values from `f32` values.
-pub trait ExtF32 {
-    /// Creates a new `Resistance` from a number of fractional milliohms (mΩ).
-    ///
-    /// The fractional part is rounded down to the nearest whole milliohm (mΩ).
+impl_resistance_from_integer!(u8);
+impl_resistance_from_integer!(u16);
+impl_resistance_from_integer!(u32);
+impl_resistance_from_integer!(u64);
+impl_resistance_from_integer!(i8);
+impl_resistance_from_integer!(i16);
+impl_resistance_from_integer!(i32);
+impl_resistance_from_integer!(i64);
+
+/// Extension trait for simple short-hands for creating `Resistance` values from floating-point values.
+pub trait FromFloat {
+    /// Creates a new `Resistance` from a fractional number of milliohms (mΩ).
     fn milli_ohms(self) -> Resistance;
 
-    /// Creates a new `Resistance` from a number of fractional ohms (Ω).
-    ///
-    /// The fractional part is rounded down to the nearest whole milliohm (mΩ).
+    /// Creates a new `Resistance` from a fractional number of ohms (Ω).
     fn ohms(self) -> Resistance;
 
-    /// Creates a new `Resistance` from a number of fractional kilohms (kΩ).
-    ///
-    /// The fractional part is rounded down to the nearest whole milliohm (mΩ).
+    /// Creates a new `Resistance` from a fractional number of kilohms (kΩ).
     fn kilo_ohms(self) -> Resistance;
 
-    /// Creates a new `Resistance` from a number of fractional megaohms (MΩ).
-    ///
-    /// The fractional part is rounded down to the nearest whole milliohm (mΩ).
+    /// Creates a new `Resistance` from a fractional number of megaohms (MΩ).
     fn mega_ohms(self) -> Resistance;
 }
 
-impl ExtF32 for f32 {
-    #[inline]
-    fn milli_ohms(self) -> Resistance {
-        assert::is_positive_value(self);
-        Resistance::from_milli_ohms(self as u32)
-    }
+macro_rules! impl_resistance_from_float {
+    ($f:ty) => {
+        impl FromFloat for $f {
+            #[inline]
+            fn milli_ohms(self) -> Resistance {
+                assert_positive_float!(self);
+                Resistance::from_milli_ohms(self as u64)
+            }
 
-    #[inline]
-    fn ohms(self) -> Resistance {
-        assert::is_positive_value(self);
-        let milliohms = helpers::checked_mul_unsigned_f32(1_000, self)
-            .expect("Overflow when converting ohms to milliohms");
-        Resistance::from_milli_ohms(milliohms)
-    }
+            #[inline]
+            fn ohms(self) -> Resistance {
+                assert_positive_float!(self);
+                let milliohms = (self as f64) * 1_000f64;
+                Resistance::from_milli_ohms(milliohms as u64)
+            }
 
-    #[inline]
-    fn kilo_ohms(self) -> Resistance {
-        assert::is_positive_value(self);
-        let milliohms = helpers::checked_mul_unsigned_f32(1_000_000, self)
-            .expect("Overflow when converting kilohms to milliohms");
-        Resistance::from_milli_ohms(milliohms)
-    }
+            #[inline]
+            fn kilo_ohms(self) -> Resistance {
+                assert_positive_float!(self);
+                let milliohms = (self as f64) * 1_000_000f64;
+                Resistance::from_milli_ohms(milliohms as u64)
+            }
 
-    #[inline]
-    fn mega_ohms(self) -> Resistance {
-        assert::is_positive_value(self);
-        let milliohms = helpers::checked_mul_unsigned_f32(1_000_000_000, self)
-            .expect("Overflow when converting megaohms to milliohms");
-        Resistance::from_milli_ohms(milliohms)
-    }
+            #[inline]
+            fn mega_ohms(self) -> Resistance {
+                assert_positive_float!(self);
+                let milliohms = (self as f64) * 1_000_000_000f64;
+                Resistance::from_milli_ohms(milliohms as u64)
+            }
+        }
+    };
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use core::cmp::Ordering;
-    use test_case::test_case;
-
-    #[test]
-    fn test_from_milli_ohms() {
-        let r = Resistance::from_milli_ohms(1_000);
-        assert_eq!(r.milli_ohms(), 1_000);
-    }
-
-    #[test]
-    fn test_milli_ohms() {
-        let r = Resistance::from_milli_ohms(1_000);
-        assert_eq!(r.milli_ohms(), 1_000);
-    }
-
-    #[test]
-    fn test_ohms() {
-        let r = Resistance::from_milli_ohms(1_000);
-        assert_eq!(r.ohms(), 1f32);
-    }
-
-    #[test]
-    fn test_kilo_ohms() {
-        let r = Resistance::from_milli_ohms(1_000_000);
-        assert_eq!(r.kilo_ohms(), 1f32);
-    }
-
-    #[test]
-    fn test_mega_ohms() {
-        let r = Resistance::from_milli_ohms(1_000_000_000);
-        assert_eq!(r.mega_ohms(), 1f32);
-    }
-
-    #[test_case(0, true; "when resistance is zero")]
-    #[test_case(1_000, false; "when resistance is not zero")]
-    fn test_is_zero(resistance: u32, expected: bool) {
-        let r = Resistance::from_milli_ohms(resistance);
-        assert_eq!(r.is_zero(), expected);
-    }
-
-    #[test]
-    fn test_zero() {
-        let r = Resistance::zero();
-        assert_eq!(r.milli_ohms(), 0);
-    }
-
-    #[test_case(0, 0, true; "when both are zero")]
-    #[test_case(1_000, 1_000, true; "when lhs equals rhs")]
-    #[test_case(1_000, 2_000, false; "when lhs does not equal rhs")]
-    fn test_eq(lhs: u32, rhs: u32, expected: bool) {
-        let lhs = Resistance::from_milli_ohms(lhs);
-        let rhs = Resistance::from_milli_ohms(rhs);
-        assert_eq!(lhs == rhs, expected);
-    }
-
-    #[test_case(1_000, 1_000, false; "when both are equal")]
-    #[test_case(2_000, 1_000, true; "when lhs is greater")]
-    #[test_case(1_000, 2_000, false; "when rhs is greater")]
-    fn test_gt(lhs: u32, rhs: u32, expected: bool) {
-        let lhs = Resistance::from_milli_ohms(lhs);
-        let rhs = Resistance::from_milli_ohms(rhs);
-        assert_eq!(lhs > rhs, expected);
-    }
-
-    #[test_case(1_000, 1_000, true; "when both are equal")]
-    #[test_case(2_000, 1_000, true; "when lhs is greater")]
-    #[test_case(1_000, 2_000, false; "when rhs is greater")]
-    fn test_gte(lhs: u32, rhs: u32, expected: bool) {
-        let lhs = Resistance::from_milli_ohms(lhs);
-        let rhs = Resistance::from_milli_ohms(rhs);
-        assert_eq!(lhs >= rhs, expected);
-    }
-
-    #[test_case(1_000, 1_000, false; "when both are equal")]
-    #[test_case(1_000, 2_000, true; "when lhs is lesser")]
-    #[test_case(2_000, 1_000, false; "when rhs is lesser")]
-    fn test_lt(lhs: u32, rhs: u32, expected: bool) {
-        let lhs = Resistance::from_milli_ohms(lhs);
-        let rhs = Resistance::from_milli_ohms(rhs);
-        assert_eq!(lhs < rhs, expected);
-    }
-
-    #[test_case(1_000, 1_000, true; "when both are equal")]
-    #[test_case(1_000, 2_000, true; "when lhs is lesser")]
-    #[test_case(2_000, 1_000, false; "when rhs is lesser")]
-    fn test_lte(lhs: u32, rhs: u32, expected: bool) {
-        let lhs = Resistance::from_milli_ohms(lhs);
-        let rhs = Resistance::from_milli_ohms(rhs);
-        assert_eq!(lhs <= rhs, expected);
-    }
-
-    #[test_case(1_000, 2_000, Ordering::Less; "when lhs is lesser")]
-    #[test_case(1_000, 1_000, Ordering::Equal; "when both are equal")]
-    #[test_case(2_000, 1_000, Ordering::Greater; "when lhs is greater")]
-    fn test_cmp(lhs: u32, rhs: u32, expected: Ordering) {
-        let lhs = Resistance::from_milli_ohms(lhs);
-        let rhs = Resistance::from_milli_ohms(rhs);
-        assert_eq!(lhs.cmp(&rhs), expected);
-    }
-
-    #[test]
-    fn test_add_operator() {
-        let lhs = Resistance::from_milli_ohms(1_000);
-        let rhs = Resistance::from_milli_ohms(2_000);
-        let expected = Resistance::from_milli_ohms(3_000);
-        assert_eq!(lhs + rhs, expected);
-    }
-
-    #[test]
-    #[should_panic(expected = "Overflow when adding resistance value")]
-    fn test_add_operator_overflow() {
-        let _r = Resistance::from_milli_ohms(u32::MAX) + Resistance::from_milli_ohms(1_000);
-    }
-
-    #[test]
-    fn test_sub_operator() {
-        let lhs = Resistance::from_milli_ohms(3_000);
-        let rhs = Resistance::from_milli_ohms(1_000);
-        let expected = Resistance::from_milli_ohms(2_000);
-        assert_eq!(lhs - rhs, expected);
-    }
-
-    #[test]
-    #[should_panic(expected = "Overflow when subtracting resistance value")]
-    fn test_sub_operator_underflow() {
-        let _r = Resistance::from_milli_ohms(0) - Resistance::from_milli_ohms(1_000);
-    }
-
-    #[test]
-    fn test_mul_operator_u32() {
-        let r = Resistance::from_milli_ohms(1_000);
-        let expected = Resistance::from_milli_ohms(10_000);
-        assert_eq!(r * 10u32, expected);
-    }
-
-    #[test]
-    #[should_panic(expected = "Overflow when multiplying resistance value")]
-    fn test_mul_operator_u32_overflow() {
-        let _r = Resistance::from_milli_ohms(u32::MAX) * 2u32;
-    }
-
-    #[test]
-    fn test_mul_operator_f32() {
-        let r = Resistance::from_milli_ohms(1_000);
-        let expected = Resistance::from_milli_ohms(2_500);
-        assert_eq!(r * 2.5f32, expected);
-    }
-
-    #[test]
-    #[should_panic(expected = "Cannot multiply resistance value by negative value")]
-    fn test_mul_operator_f32_negative() {
-        let _r = Resistance::from_milli_ohms(1_000) * -1f32;
-    }
-
-    #[test_case(f32::INFINITY; "positive infinity")]
-    #[test_case(f32::NEG_INFINITY; "negative infinity")]
-    #[should_panic(expected = "Cannot multiply resistance value by infinity")]
-    fn test_mul_operator_f32_infinity(infinity: f32) {
-        let _r = Resistance::from_milli_ohms(1_000) * infinity;
-    }
-
-    #[test_case(f32::NAN; "NaN")]
-    #[should_panic(expected = "Cannot multiply resistance value by NaN")]
-    fn test_mul_operator_f32_nan(nan: f32) {
-        let _r = Resistance::from_milli_ohms(1_000) * nan;
-    }
-
-    #[test]
-    fn test_div_operator_u32() {
-        let r = Resistance::from_milli_ohms(10_000);
-        let expected = Resistance::from_milli_ohms(1_000);
-        assert_eq!(r / 10u32, expected);
-    }
-
-    #[test]
-    #[should_panic(expected = "Cannot divide resistance value by zero")]
-    fn test_div_operator_u32_by_zero() {
-        let _r = Resistance::from_milli_ohms(1_000) / 0;
-    }
-
-    #[test]
-    fn test_div_operator_f32() {
-        let r = Resistance::from_milli_ohms(2_500);
-        let expected = Resistance::from_milli_ohms(1_000);
-        assert_eq!(r / 2.5f32, expected);
-    }
-
-    #[test]
-    #[should_panic(expected = "Cannot divide resistance value by zero")]
-    fn test_div_operator_f32_by_zero() {
-        let _r = Resistance::from_milli_ohms(1_000) / 0f32;
-    }
-
-    #[test]
-    #[should_panic(expected = "Cannot divide resistance value by negative value")]
-    fn test_div_operator_f32_negative() {
-        let _r = Resistance::from_milli_ohms(1_000) / -1f32;
-    }
-
-    #[test_case(f32::INFINITY; "positive infinity")]
-    #[test_case(f32::NEG_INFINITY; "negative infinity")]
-    #[should_panic(expected = "Cannot divide resistance value by infinity")]
-    fn test_div_operator_f32_infinity(infinity: f32) {
-        let _r = Resistance::from_milli_ohms(1_000) / infinity;
-    }
-
-    #[test_case(f32::NAN; "NaN")]
-    #[should_panic(expected = "Cannot divide resistance value by NaN")]
-    fn test_div_operator_f32_nan(nan: f32) {
-        let _r = Resistance::from_milli_ohms(1_000) / nan;
-    }
-
-    #[test]
-    fn test_milli_ohms_u32() {
-        let r = 1_000u32.milli_ohms();
-        assert_eq!(r.milli_ohms(), 1_000u32);
-    }
-
-    #[test]
-    fn test_ohms_u32() {
-        let r = 1_000u32.ohms();
-        assert_eq!(r.ohms(), 1_000f32);
-    }
-
-    #[test]
-    fn test_kilo_ohms_u32() {
-        let r = 1_000u32.kilo_ohms();
-        assert_eq!(r.kilo_ohms(), 1_000f32);
-    }
-
-    #[test]
-    fn test_mega_ohms_u32() {
-        let r = 2u32.mega_ohms();
-        assert_eq!(r.mega_ohms(), 2f32);
-    }
-
-    #[test]
-    #[should_panic(expected = "Overflow when converting megaohms to milliohms")]
-    fn test_mega_ohms_u32_overflow() {
-        let _r = 10u32.mega_ohms();
-    }
-
-    #[test]
-    fn test_milli_ohms_f32() {
-        let r = 1_000f32.milli_ohms();
-        assert_eq!(r.milli_ohms(), 1_000u32);
-    }
-
-    #[test]
-    #[should_panic(expected = "Value cannot be negative")]
-    fn test_milli_ohms_f32_negative() {
-        let _r = (-1f32).milli_ohms();
-    }
-
-    #[test]
-    fn test_ohms_f32() {
-        let r = 1_000f32.ohms();
-        assert_eq!(r.ohms(), 1_000f32);
-    }
-
-    #[test]
-    #[should_panic(expected = "Value cannot be negative")]
-    fn test_ohms_f32_negative() {
-        let _r = (-1f32).ohms();
-    }
-
-    #[test]
-    fn test_kilo_ohms_f32() {
-        let r = 1_000f32.kilo_ohms();
-        assert_eq!(r.kilo_ohms(), 1_000f32);
-    }
-
-    #[test]
-    #[should_panic(expected = "Value cannot be negative")]
-    fn test_kilo_ohms_f32_negative() {
-        let _r = (-1f32).kilo_ohms();
-    }
-
-    #[test]
-    fn test_mega_ohms_f32() {
-        let r = 2f32.mega_ohms();
-        assert_eq!(r.mega_ohms(), 2f32);
-    }
-
-    #[test]
-    #[should_panic(expected = "Value cannot be negative")]
-    fn test_mega_ohms_f32_negative() {
-        let _r = (-1f32).mega_ohms();
-    }
-
-    #[test]
-    #[should_panic(expected = "Overflow when converting megaohms to milliohms")]
-    fn test_mega_ohms_f32_overflow() {
-        let _r = 10f32.mega_ohms();
-    }
-}
+impl_resistance_from_float!(f32);
+impl_resistance_from_float!(f64);